Centralize host Wasm boundary helpers
