+++ host/memory.mjs
@@ -1,3 +1,48 @@
+export function globalValue(value) {
+  return value instanceof WebAssembly.Global ? value.value : value;
+}
+
+export function errorMessage(error) {
+  return error instanceof Error ? error.message : String(error);
+}
+
+export function promisingWasmExport(fn, receiver = undefined) {
+  return typeof WebAssembly.promising === "function"
+    ? WebAssembly.promising(fn)
+    : fn.bind(receiver);
+}
+
+export function normalizedRowCap(value, fallback) {
+  const numeric = Number(value);
+
+  return Number.isFinite(numeric) && numeric >= 0
+    ? Math.floor(numeric)
+    : fallback;
+}
+
+export function normalizedPositiveInteger(value, fallback) {
+  const numeric = Number(value);
+
+  return Number.isFinite(numeric) && numeric > 0
+    ? Math.floor(numeric)
+    : fallback;
+}
+
+export function numericSize(value) {
+  if (value === undefined || value === null) {
+    return null;
+  }
+
+  const numeric = Number(value);
+  return Number.isFinite(numeric) && numeric >= 0 ? numeric : null;
+}
+
+export function wasmNumber(value, fallback) {
+  const numeric = Number(globalValue(value));
+
+  return Number.isFinite(numeric) ? numeric : fallback;
+}
+
 export function u64ToNumber(value) {
   const number = typeof value === "bigint" ? Number(value) : value;
